+++ src/bitter/msb0.hpp
@@ -40,8 +40,8 @@ struct msb0
     /// field sizes once we reach the index we are interested in we
     /// sum the remaining sizes.
     ///
-    template<uint32_t Index, uint32_t... Sizes>
-    static uint32_t field_offset()
+    template<uint32_t Index, std::size_t... Sizes>
+    static std::size_t field_offset()
     {
         return count_to_field_offset<Index, 0, Sizes...>();
     }
@@ -50,11 +50,11 @@ private:
 
     template
     <
-        uint32_t Index,
-        uint32_t Counter,
-        uint32_t Size0
+        std::size_t Index,
+        std::size_t Counter,
+        std::size_t Size0
     >
-    static uint32_t count_to_field_offset()
+    static std::size_t count_to_field_offset()
     {
         static_assert(Index == Counter, "");
         return 0;
@@ -62,13 +62,13 @@ private:
 
     template
     <
-        uint32_t Index,
-        uint32_t Counter,
-        uint32_t Size0,
-        uint32_t Size1,
-        uint32_t... Sizes
+        std::size_t Index,
+        std::size_t Counter,
+        std::size_t Size0,
+        std::size_t Size1,
+        std::size_t... Sizes
     >
-    static uint32_t count_to_field_offset()
+    static std::size_t count_to_field_offset()
     {
         if (Index == Counter)
         {